bot_requirements.txt
python-telegram-bot==20.1
supabase==2.3.1
Pillow==10.0.0
python-dotenv==1.0.0
aiohttp==3.8.5
Flask==2.3.2